P1_P2.py created, former pipeline.py functions.
pipeline.py has functions moved to p1_p2.py.
helper.py created to include general use functions.
